Adding back the scores function
After some learning (if you can call it that) I have added back the
function to write scores to a file. Per default the script is located in
/usr/local/share and writes to $HOME/.2048; To change the script
location you have to change the Makefile SCRIPT variable and the
SCORES_SCRIPT #define in main.h and to change the scores file location
change the variable in scores.sh ( there is only one ).