Change signature of function spawn
The signature of the function spawn is changed from it's original, that
is void spawn(board_t *b), to a new one, that is int spawn(board_t *b,
int callc). This allows us to introduce the concept of a call limit to
spawn. Since spawn calls spawn if it could not spawn a new tile at the
generated position, it is theoretically possible that due to unforseen
circumstances we are in a state that is technically game over, but still
called spawn. If that were to be the case, in the old system we would
have called spawn until the OOM-Reaper said hello; With the new system
in place if the call count reaches NUM_CELLS * NUM_COLUMNS we return 0
and exit the game loop, simiar to exiting after game over or after
pressing r/q.

That is basically it. This also allows us to combine the first call of
spawn with the assignment of stop, which is neat, I think.